game_helper/requirements.txt
Pillow
pyautogui
pytesseract
keyboard